pass op_ctx to inner_matches_operator; rename json_ops modules
