+++ packages/cw-jsonfilter/src/json_ops/mod.rs
@@ -1,7 +1,7 @@
-pub mod collection;
-pub mod comparison;
-pub mod existence;
-pub mod logical;
+pub mod op_collection;
+pub mod op_comparison;
+pub mod op_existence;
+pub mod op_logical;
+pub mod op_transform;
+pub mod op_type;
 pub mod operator;
-pub mod transform;
-pub mod types;
