+++ packages/cw-jsonfilter/src/json_ops/op_type.rs
